refactor: rename search-page file to search-page-pure
