+++ src/pages/search-page/search-page-pure.jsx
